Django/requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1.1
django-debug-toolbar==3.6.0
djangorestframework==3.13.1
docopt==0.6.2
idna==3.4
pipreqs==0.4.11
psycopg2==2.9.3
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
yarg==0.1.9
